requirements.txt
aiodns==3.4.0
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
frozenlist==1.6.0
greenlet==3.2.2
idna==3.10
loguru==0.7.3
magic-filter==1.0.12
multidict==6.4.4
propcache==0.3.1
psycopg==3.2.9
psycopg-binary==3.2.9
pycares==4.8.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvloop==0.21.0
yarl==1.20.0
